+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/Internal.kt
@@ -1,5 +1,0 @@
-package com.huanshankeji
-
-@RequiresOptIn("This API is internal in this project.")
-@Retention(AnnotationRetention.BINARY)
-annotation class InternalApi