+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/SnapshotAndReleaseVersions.kt
@@ -1,6 +1,0 @@
-package com.huanshankeji
-
-import org.gradle.api.Project
-
-fun Project.isSnapshotVersion() =
-    version.toString().endsWith("SNAPSHOT")